+++ .storybook/preview.ts
@@ -1,6 +1,6 @@
 // Tokens
-import '../packages/tokens/dist/web/core.css';
-import '../packages/tokens/dist/web/components.css';
+import '../packages/tokens/dist/css/core.css';
+import '../packages/tokens/dist/css/components.css';
 
 // Base
 import '../packages/ui/dist/css/fonts.css';